+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.3-pre
 * Add list-dependencies, list-devdependencies, list-alldependencies commands
+* Add sort-dependencies, sort-devdependencies, sort-alldependencies commands
 
 ### 0.0.2
 * Add range version support to dependencies upgrade